requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.24.2,<0.25.0
djangorestframework-simplejwt>=5.2.0,<5.3